% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nmap}
\alias{nmap}
\title{nmap}
\usage{
nmap(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run nmap
}
\examples{
library(outsider)
nmap <- module_import('nmap', repo = 'hrbrmstr/om..nmap')
nmap('-h')  # or --help or whichever argument prints help
}
